refactor(ppiflow): remove the previous draft and related tests
