--- tests/workflow/test_ppiflow_workflow_inputs.py
@@ -1,80 +1,0 @@
-"""Tests for PPIFlow workflow input staging helpers."""
-
-# ruff: noqa: D103
-
-from pathlib import Path
-
-import pytest
-
-from biomodals.workflow.ppiflow_workflow import (
-    _collect_local_inputs,
-    build_local_output_path,
-)
-
-
-def test_collect_local_inputs_rejects_sanitized_name_collisions(
-    tmp_path: Path,
-) -> None:
-    task_path = tmp_path / "task.yaml"
-    task_path.write_text("task: demo\n", encoding="utf-8")
-    nested = tmp_path / "a"
-    nested.mkdir()
-    nested.joinpath("b.pdb").write_text("nested", encoding="utf-8")
-    tmp_path.joinpath("a_b.pdb").write_text("flat", encoding="utf-8")
-
-    with pytest.raises(ValueError, match="Duplicate staged input path: a_b.pdb"):
-        _collect_local_inputs(
-            {
-                "task": {
-                    "nested_path": "a/b.pdb",
-                    "flat_path": "a_b.pdb",
-                }
-            },
-            task_path,
-        )
-
-
-def test_collect_local_inputs_returns_sanitized_keys(tmp_path: Path) -> None:
-    task_path = tmp_path / "task.yaml"
-    task_path.write_text("task: demo\n", encoding="utf-8")
-    nested = tmp_path / "a"
-    nested.mkdir()
-    nested.joinpath("b.pdb").write_text("nested", encoding="utf-8")
-
-    staged = _collect_local_inputs({"task": {"input_path": "a/b.pdb"}}, task_path)
-
-    assert staged == [("a_b.pdb", b"nested")]
-
-
-def test_collect_local_inputs_stages_steps_yaml_references(
-    tmp_path: Path,
-) -> None:
-    task_path = tmp_path / "task.yaml"
-    task_path.write_text("task: demo\n", encoding="utf-8")
-    steps_dir = tmp_path / "steps"
-    steps_dir.mkdir()
-    steps_path = steps_dir / "steps.yaml"
-    steps_path.write_text("PPIFlowStep: {}\n", encoding="utf-8")
-    steps_dir.joinpath("constraints.csv").write_text(
-        "chain,pos\nA,1\n", encoding="utf-8"
-    )
-
-    staged = _collect_local_inputs(
-        {"task": {"input_path": "input.pdb"}},
-        task_path,
-        {"PPIFlowStep": {"constraint_csv": "constraints.csv"}},
-        steps_path,
-    )
-
-    assert staged == [("constraints.csv", b"chain,pos\nA,1\n")]
-
-
-def test_build_local_output_path_allows_blank_optional_parts(tmp_path: Path) -> None:
-    out_path = build_local_output_path(
-        tmp_path,
-        run_name="demo",
-        prefix=" ",
-        suffix="",
-    )
-
-    assert out_path == tmp_path / "demo.tar.zst"
